PHP 8 and strange configuration issue fixes
thanks to colinax and bernd

+++ wbce/framework/initialize.php
@@ -180,7 +180,7 @@ switch (true) {
 
     case (ER_LEVEL == "E0"):
         // system default (php.ini)
-        error_reporting(ini_get('error_reporting'));
+         error_reporting((int)ini_get('error_reporting'));
         break;
 
     case (ER_LEVEL == "E1"):
